# src/main/resources/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost
    description: Generated server url
...
paths:
...
  "/chattingRoom":
    get:
      tags:
        - Chat-Controller
      summary: 현재 채팅방
      description: 저장된 쿠키를 기반으로 사용자가 현재 참여 중인 채팅방을 검색합니다.
      operationId: chattingRoom_1
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: object
    post:
      tags:
        - Chat-Controller
      summary: 채팅방 만들기
      description: 새 채팅방을 만들고 사용 가능한 방 목록에 추가합니다.
      operationId: chattingRoom
      parameters:
        - name: roomName
          in: query
          required: true
          schema:
            type: string
        - name: nickname
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: object
  "/cart":
    get:
      tags:
        - Cart-Controller
      summary: 장바구니 목록 조회
      description: 사용자의 장바구니에 담긴 상품 목록을 조회합니다.
      operationId: orderHist_2
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
    post:
      tags:
        - Cart-Controller
      summary: 장바구니에 상품 추가
      description: 장바구니에 새로운 상품을 추가합니다.
      operationId: order_1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CartItemDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/cart/orders":
    post:
      tags:
        - Cart-Controller
      summary: 장바구니 상품 주문
      description: 장바구니에 담긴 상품을 주문합니다.
      operationId: orderCartItem
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CartOrderDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/chattingRoom-exit":
    patch:
      tags:
        - Chat-Controller
      summary: 채팅방 나가기
      description: 사용자가 채팅방을 나가고 방의 사용자 목록을 그에 따라 업데이트합니다.
      operationId: ExitChattingRoom
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ChattingRoomExitDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: object
  "/cartItem/{cartItemId}":
    delete:
      tags:
        - Cart-Controller
      summary: 장바구니 상품 삭제
      description: 장바구니에서 특정 상품을 삭제합니다.
      operationId: deleteCartItem
      parameters:
        - name: cartItemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
    patch:
      tags:
        - Cart-Controller
      summary: 장바구니 상품 수량 변경
      description: 장바구니에 담긴 특정 상품의 수량을 변경합니다.
      operationId: updateCartItem
      parameters:
        - name: cartItemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: count
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/orders/{page}":
    get:
      tags:
        - Order Controller
      summary: 주문 내역 조회
      description: 사용자의 주문 내역을 페이지네이션으로 조회합니다.
      operationId: orderHist_1
      parameters:
        - name: page
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/chattingRoom-enter":
    get:
      tags:
        - Chat-Controller
      summary: 채팅방 들어가기
      description: 사용자가 방 번호를 사용하여 채팅방에 입장할 수 있습니다.
      operationId: EnterChattingRoom
      parameters:
        - name: roomNumber
          in: query
          required: true
          schema:
            type: string
        - name: nickname
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: object
...
components:
  schemas:
...
    CartItemDto:
      required:
        - itemId
      type: object
      properties:
        itemId:
          type: integer
          format: int64
        count:
          minimum: 1
          type: integer
          format: int32
    CartOrderDto:
      type: object
      properties:
        cartItemId:
          type: integer
          format: int64
        cartOrderDtoList:
          type: array
          items:
            "$ref": "#/components/schemas/CartOrderDto"
        impUid:
          type: string
...
    ChattingRoomExitDto:
      type: object
      properties:
        roomNumber:
          type: string
        nickname:
          type: string
...